+++ changelog.md
@@ -3,6 +3,7 @@
 * Fixed odd resolution filter behavior
 * Fixed identical images but different settings colliding
 * Added `enableBundledPaintings` config to disable bundled paintings
+* Added `loadPaintings` config to disable downloading and rendering paintings
 
 # 0.7.7